requirements.txt
torch
torchvision
opencv-python
numpy
pillow
python-dotenv
pandas